Update several features related to the operation and customer ease of use for the Cyberphysical hub:
1. Eased the restrictions on what fields needed
to be filled out when adding or editing a device.
Only the device name is required now.

2. Cleaned up the verbage of the device connection
indicators on the device page. Connection status
used to reflect registration status which created
confusion.  Now it is only associated with actual
tunnel connection status.

3.  A delete button has now been added on the
individual device page in addition to the list of
all devices

4. The main page has received a UI overhaul and
includes the following:
    - Changed the device boxes from big, bulky
    squares to more manageable tabs to fit more
    - Replaced the words for each button function
    to icons in order from left to right:
        Remote
        Ping
        Camera a
        Camera b
        Restart *Coming Soon*